backend-api/requirements.txt
# SENTINEL API — Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.30.0
geoalchemy2==0.15.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
httpx==0.27.0
python-socketio==5.11.0
boto3==1.35.0
Pillow==10.4.0
pymilvus==2.4.0
